move functions from common data file to rates data file and update path imports
