+++ frontend/pages/rates.py
@@ -16,10 +16,13 @@ from pages.data.data import (
     clients_names,
     to_rate,
     months_start,
+    username,
+)
+
+from pages.data.rates import (
     rates_by_user_client_date,
     rate_active_by_user_client,
     rate_update,
-    username,
 )
 
 from pages.utils import month_start_to_str, far_date, date_str_to_date
